fix(head): make useHead DOM mutation a no-op in environments without a DOM
useHead mutated the DOM unconditionally in its effect, so any environment
with neither a DOM nor an SSR HeadProvider collector (edge runtimes, node
test environments) crashed with `ReferenceError: document is not defined`.

Extract the DOM-mutation branch into an applyHeadTagsToDOM helper whose
first line is an environment guard (`typeof document === "undefined"` ->
no-op returning no managed elements). Browser behaviour is unchanged and
the package's 100% coverage thresholds still pass; two node-environment
tests cover the guard, including renderToString without a provider.

Ported from the still-valid half of #663.
